requirements.txt
bs4~=0.0.1
beautifulsoup4~=4.9.3
Pillow~=8.2.0
requests~=2.25.1